refactor: eventListContext 모듈 분리, getEventList test API module 구현
